Remove unused code
